backend/requirements.txt
Flask==2.3.3
Flask-CORS==3.0.10
Flask-JWT-Extended==4.5.2
Flask-SocketIO==5.3.6
werkzeug==2.3.7
supabase
langchain==0.2.16
langchain-groq==0.1.9
langsmith==0.1.115
openai==0.27.0
python-dotenv==1.0.0
gunicorn==20.1.0
tenacity==8.5.0
